requirements.txt
torch
torchvision
diffusers
transformers
datasets
accelerate
peft
pillow
